Add DOI validator to publication
An earlier version had this validation as part of the dara extension as a validator for the form. However, there was a desire to have the valdiation happen only on publication. Inorder to differentiate between different states the function needed to called 'package_show'. However, 'package_show' does validation which caused a crash from an inifinite loop. The validation is now part of the controller to avoid that.